added unit-tests with mocks for UserService class, added constructors in UserDAO and UserService

+++ src/main/java/org/example/UserService.java
@@ -1,74 +1,54 @@
 package org.example;
 
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Scanner;
 
 public class UserService {
 
-    static final Logger logger = LoggerFactory.getLogger(UserService.class);
-    private static UserDAO userDAO = new UserDAO();
-    static Scanner scanner = new Scanner(System.in);
+    private final Logger logger;
+    private UserDAO userDAO;
 
-    public void menu() {
-        logger.info("Меню:");
-        System.out.println("1. Создать пользователя");
-        System.out.println("2. Найти пользователя");
-        System.out.println("3. Обновить данные пользователя");
-        System.out.println("4. Удалить пользователя");
-        System.out.println("5. Завершить работу");
+    public UserService(Logger logger, UserDAO userDAO) {
+        this.logger = logger;
+        this.userDAO = userDAO;
     }
 
-    public void createUser() {
-        User user = new User();
-        logger.info("Введите имя нового пользователя:");
-        user.setName(scanner.nextLine());
-        logger.info("Введите электронную почту:");
-        user.setEmail(scanner.nextLine());
-        logger.info("Введите возраст:");
-        try {
-            user.setAge(scanner.nextInt());
-        } catch (IllegalArgumentException e) {
-            logger.warn("Ошибка ввода: необходимо ввести корректное значение возраста!");
-            return;
-        }
+    public void createUser(User user) {
         Integer id = userDAO.create(user);
         if (id != 0) {
             logger.info("Пользователь с ID " + id + " успешно создан!");
+        } else {
+            logger.error("Невозможно создать пользователя: данный email уже используется!");
         }
     }
 
-    public void readUser() {
-        logger.info("Введите ID пользователя:");
-        User user = userDAO.read(scanner.nextInt());
+    public User readUser(int id) {
+        User user = userDAO.read(id);
+        if(user == null) {
+            logger.error("Пользователя с ID " + id + " не существует!");
+            return null;
+        }
         logger.info("Имя пользователя: " + user.getName());
         logger.info("Электронная почта: " + user.getEmail());
         logger.info("Возраст: " + user.getAge());
         logger.info("Дата создания записи: " + user.getCreatedAt());
+        return user;
     }
 
-    public void updateUser() {
-        User user = new User();
-        logger.info("Введите ID пользователя:");
-        user.setId(scanner.nextInt());
-        scanner.nextLine();
-        logger.info("Введите имя:");
-        user.setName(scanner.nextLine());
-        logger.info("Введите электронную почту:");
-        user.setEmail(scanner.nextLine());
-        logger.info("Введите возраст:");
-        user.setAge(scanner.nextInt());
+    public void updateUser(User user) {
         if (userDAO.update(user)) {
             logger.info("Данные успешно обновлены!");
+        } else {
+            logger.error("Невозможно обновить данные: данный email уже используется!");
         }
     }
 
-    public void deleteUser() {
-        logger.info("Введите ID пользователя");
-        Integer id = scanner.nextInt();
+    public void deleteUser(int id) {
         User user = userDAO.read(id);
-        userDAO.delete(user);
-        logger.info("Пользователь с ID " + id + " успешно удален!");
+        if(user != null) {
+            userDAO.delete(user);
+            logger.info("Пользователь с ID " + id + " успешно удален!");
+        } else {
+            logger.error("Пользователя с ID " + id + " не существует!");
+        }
     }
 }
